refactor(domain): `WinningLotto`를 `entity`에서 `domain` 패키지로 변경
- `WinningLotto`는 DB 엔티티가 아닌 핵심 도메인 객체이므로 `domain` 패키지로 이동
- 비즈니스 로직을 포함하는 객체이므로 `entity`보다 `domain`이 더 적절
- ORM 도입 시 `entity` 패키지를 JPA 엔티티 관리 용도로 활용 가능
- 패키지 변경으로 유지보수성과 코드 명확성 향상
