+++ src/main/java/lotto/model/domain/WinningLotto.java
@@ -1,4 +1,4 @@
-package lotto.model.entity;
+package lotto.model.domain;
 
 import java.util.List;
 import java.util.Set;
